perf: realizando ajustes de performance no banco e inserindo a verificacao do tenant nos metodos do service de servico

+++ backend/src/main/java/cloud/zenixapp/zenix/services/ServicoService.java
@@ -1,7 +1,10 @@
 package cloud.zenixapp.zenix.services;
 
+import cloud.zenixapp.zenix.configs.TenantContext;
+import cloud.zenixapp.zenix.configs.exceptions.ExistsException;
 import cloud.zenixapp.zenix.configs.exceptions.NotFoundException;
 import cloud.zenixapp.zenix.configs.exceptions.ServicoExcluidoException;
+import cloud.zenixapp.zenix.configs.exceptions.OptimisticException;
 import cloud.zenixapp.zenix.configs.mappers.ServicoMapper;
 import cloud.zenixapp.zenix.models.dtos.requests.ServicoRequestDTO;
 import cloud.zenixapp.zenix.models.dtos.responses.ServicoResponseDTO;
@@ -9,12 +12,12 @@ import cloud.zenixapp.zenix.models.dtos.responses.SuccessDeleteServicoResponseDT
 import cloud.zenixapp.zenix.models.dtos.responses.SuccessServicoResponseDTO;
 import cloud.zenixapp.zenix.models.entities.Servicos;
 import cloud.zenixapp.zenix.repositories.ServicoRepository;
+import jakarta.persistence.OptimisticLockException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -28,44 +31,69 @@ public class ServicoService {
 
     @Transactional
     public SuccessServicoResponseDTO inserirServico(ServicoRequestDTO servicoRequestDTO){
-        Servicos servico = servicoRepository.save(servicoMapper.toServicos(servicoRequestDTO));
+        String tenantId = TenantContext.getTenantId();
+        if (servicoRepository.existsServicoByServicoAndTenantId(servicoRequestDTO.servico(), tenantId)){
+            throw new ExistsException("Serviço já existe!");
+
+        }
+
+        Servicos servico = servicoMapper.toServicos(servicoRequestDTO);
+        servico.setTenantId(tenantId);
+
+        servicoRepository.save(servico);
 
         return new SuccessServicoResponseDTO(
                 HttpStatus.OK.value(),
                 "Serviço inserido com sucesso",
-                servico
+                servicoMapper.toServicoDTO(servico)
         );
     }
 
     public List<ServicoResponseDTO> buscarTodosServicos(){
-        return servicoMapper.toListServicos(servicoRepository.findAll());
+        return servicoMapper.toListServicos(servicoRepository.findAll(TenantContext.getTenantId()));
     }
 
     @Transactional
     public SuccessServicoResponseDTO atualizarServico(ServicoRequestDTO servicoRequestDTO, String id){
-        return servicoRepository.findById(id)
+        return servicoRepository.findById(id, TenantContext.getTenantId())
                 .map(servico -> {
                     if (servico.getStatus() == -1){
                         throw new ServicoExcluidoException("Serviço foi excluído!");
                     }
 
-                    servico.setUpdatedAt(LocalDateTime.now());
-                    servicoMapper.atualizarServico(servico, servicoRequestDTO);
-                    servicoRepository.save(servico);
+                    try {
+                        servicoMapper.atualizarServico(servico, servicoRequestDTO);
+                        servicoRepository.save(servico);
+
+                    } catch (OptimisticLockException e){
+                        throw new OptimisticException("Erro ao atualizar: " + e + ", tente novamente!");
+
+                    }
+
 
                     return new SuccessServicoResponseDTO(
                             HttpStatus.OK.value(),
                             "Serviço atualizado com sucesso",
-                            servico
+                            servicoMapper.toServicoDTO(servico)
                     );
                 }).orElseThrow(() -> new NotFoundException("Serviço não encontrado"));
     }
 
     @Transactional
     public SuccessDeleteServicoResponseDTO deletarServico(String id) {
-        return servicoRepository.findById(id)
+        return servicoRepository.findById(id, TenantContext.getTenantId())
                 .map(servico -> {
-                    servicoRepository.deleteById(id);
+                    if (servico.getStatus() == -1){
+                        throw new ServicoExcluidoException("Serviço foi excluído!");
+                    }
+
+                    try {
+                        servicoRepository.deleteById(id);
+
+                    } catch (OptimisticLockException e){
+                        throw new OptimisticException("Erro ao deletar: " + e + ", tente novamente!");
+
+                    }
 
                     return new SuccessDeleteServicoResponseDTO(
                             HttpStatus.OK.value(),
